remove redundant child reducer, fix tests, add disable loading test
